# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LifeLoop API",
    "version": "1.0.0",
    "description": "The backend API powering LifeLoop — scanning, AI recommendations, community, gamification, and notifications. All endpoints except registration/login require a valid session cookie (obtained by logging in)."
  },
  "servers": [{ "url": "/api", "description": "Local server" }],
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "lifeloop_token" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "INVALID_NAME" },
          "message": { "type": "string", "example": "Please tell us what the item is." }
        }
      },
...
      "Recommendation": {
        "type": "object",
        "properties": {
          "action": { "type": "string", "enum": ["keep", "repair", "reuse", "resell", "donate", "recycle"] },
          "score": { "type": "integer", "minimum": 0, "maximum": 100 },
          "confidence": { "type": "string", "enum": ["low", "medium", "high"] },
          "explanation": { "type": "string" }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "category": { "type": "string", "enum": ["clothing", "electronics", "tech", "furniture", "home", "other"] },
          "condition": { "type": "string", "enum": ["excellent", "good", "fair", "worn", "poor", "broken"] },
          "ageYears": { "type": "number", "nullable": true },
          "images": { "type": "array", "items": { "type": "object", "properties": { "url": { "type": "string" } } } },
          "identification": { "type": "object" },
          "recommendations": { "type": "array", "items": { "$ref": "#/components/schemas/Recommendation" } },
          "primaryAction": { "type": "string" },
          "lifePotential": { "type": "integer" },
          "note": { "type": "string" },
          "source": { "type": "string", "enum": ["builtin"] },
          "userAction": { "type": "string", "nullable": true },
          "saved": { "type": "boolean" },
          "notes": { "type": "string" }
        }
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/items/scan": {
      "post": {
        "summary": "Scan a new item and get ranked recommendations",
        "description": "Multipart form. Fields: name (required), category, condition, ageYears, brand, model, material, runOcr ('true' to extract text from the first photo via OCR), photos (up to 6 image files).",
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object" } } } },
        "responses": {
          "201": { "description": "Item created", "content": { "application/json": { "schema": { "type": "object", "properties": { "item": { "$ref": "#/components/schemas/Item" } } } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/items/scan-code": {
      "post": {
        "summary": "Decode a QR code from a photo",
        "description": "Multipart form field: codeImage (single image). Returns the decoded text and, if BARCODE_LOOKUP_API_URL is configured, a product lookup result. 1D barcodes (UPC/EAN) are not supported.",
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object" } } } },
        "responses": {
          "200": { "description": "Code decoded", "content": { "application/json": { "schema": { "type": "object", "properties": { "code": { "type": "string" }, "product": { "nullable": true } } } } } },
          "404": { "description": "No code found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/items": {
      "get": {
        "summary": "List/search/filter/paginate your scanned items",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "condition", "in": "query", "schema": { "type": "string" } },
          { "name": "action", "in": "query", "schema": { "type": "string" } },
          { "name": "saved", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "name": "completed", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest", "oldest", "score_desc", "score_asc"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Paginated item list" } }
      }
    },
...
    "/items/{id}": {
      "get": { "summary": "Get a single item", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Item" }, "404": { "description": "Not found" } } },
      "patch": { "summary": "Correct fields, record an actual action taken, save/unsave, or add notes", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Updated item" } } },
      "delete": { "summary": "Delete an item and its photos", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/items/{id}/assistant/{type}": {
      "get": {
        "summary": "Generate (or fetch cached) repair/reuse/resell/donate/recycle guide",
        "description": "Returns a cached guide immediately, or enqueues a background job (poll /jobs/{jobId}) if not yet generated.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "type", "in": "path", "required": true, "schema": { "type": "string", "enum": ["repair", "reuse", "resell", "donate", "recycle"] } },
          { "name": "regenerate", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
        ],
        "responses": {
          "200": { "description": "Cached guide returned immediately" },
          "202": { "description": "Guide generation queued", "content": { "application/json": { "schema": { "type": "object", "properties": { "jobId": { "type": "string" }, "status": { "type": "string" } } } } } }
        }
      }
    },
...
  }
}